+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.4.7] - 2025-01-20
+### Changed
+- ClassWeb search will search `LCSH (w/names)` to do a combined search.
+- Ability to keyword search subject by appending `?` to the search term
+
 ## [1.4.6] - 2025-01-16
 ### Added
 - Added support for automatically populating 4xx from the MARC key if present.
 - NAR will strip trailing period if at end of string and previous char is lowercase.
+- If part of a literal field's text is highlighted (selected) and scriptshifter is envoked it will only operate on the selected text.
 
 ## [1.4.5] - 2025-12-16
 ### Added